requirements.txt
bs4
requests
PySocks